Refactor application settings
