+++ StudentUsos/Features/Settings/Views/Subpages/ApplicationSubpage.xaml
@@ -1,0 +1,30 @@
+﻿<?xml version="1.0" encoding="utf-8" ?>
+<controls:CustomContentPageNotAnimated
+    x:Class="StudentUsos.Features.Settings.Views.Subpages.ApplicationSubpage"
+    xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+    xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    xmlns:controls="clr-namespace:StudentUsos.Controls"
+    xmlns:strings="clr-namespace:StudentUsos.Resources.LocalizedStrings"
+    xmlns:toolkit="http://schemas.microsoft.com/dotnet/2022/maui/toolkit"
+    xmlns:subpages="clr-namespace:StudentUsos.Features.Settings.Views.Subpages"
+    x:DataType="subpages:ApplicationSubpageViewModel">
+
+    <StackLayout Margin="30">
+
+        <Label FontSize="14" HorizontalTextAlignment="Center" Text="{x:Static strings:LocalizedStrings.Language}" />
+        <Border Margin="0,5,0,10" Padding="0">
+            <Button Text="{Binding CurrentLanguageName}" Padding="15"
+                        Command="{Binding LanguageButtonClickedCommand}"/>
+        </Border>
+
+        <!--<Label FontSize="14" HorizontalTextAlignment="Center" Text="{x:Static strings:LocalizedStrings.SettingsPage_AppInfo}" />-->
+        <Border Margin="0,5,0,20" Padding="0">
+            <StackLayout Spacing="10">
+                    <Button Text="{x:Static strings:LocalizedStrings.SettingsPage_AppInfo}" Padding="15"
+                            Clicked="AppInfoButton_Clicked"/>
+            </StackLayout>
+        </Border>
+
+    </StackLayout>
+
+</controls:CustomContentPageNotAnimated>
