+++ StudentUsos/Features/Settings/Views/Subpages/ApplicationSubpage.xaml.cs
@@ -1,0 +1,19 @@
+﻿using StudentUsos.Controls;
+
+namespace StudentUsos.Features.Settings.Views.Subpages;
+
+public partial class ApplicationSubpage : CustomContentPageNotAnimated
+{
+    INavigationService navigationService;
+    public ApplicationSubpage(ApplicationSubpageViewModel applicationSubpageViewModel, INavigationService navigationService)
+    {
+        InitializeComponent();
+        BindingContext = applicationSubpageViewModel;
+        this.navigationService = navigationService;
+    }
+
+    private async void AppInfoButton_Clicked(object sender, EventArgs e)
+    {
+        await navigationService.PushAsync<AppInfoPage>();
+    }
+}
