+++ StudentUsos/Views/BasicPopups/PickFromListPopup.xaml
@@ -10,16 +10,25 @@
 
         <Border StrokeShape="RoundRectangle 20" Margin="30" Padding="30,30,30,0" VerticalOptions="CenterAndExpand" MaximumHeightRequest="500">
             <StackLayout HorizontalOptions="CenterAndExpand" VerticalOptions="CenterAndExpand">
-                <Label x:Name="titleLabel" Text="{Binding PopupTitle,Source={x:Reference self}}" Margin="0,0,0,20" FontAttributes="Bold" FontSize="20" HorizontalTextAlignment="Center" />
+                <Label x:Name="titleLabel" Text="{Binding PopupTitle,Source={x:Reference self}}" 
+                       Margin="0,0,0,20" FontAttributes="Bold" 
+                       FontSize="20" HorizontalTextAlignment="Center" />
 
-                <StackLayout toolkit:StateContainer.CurrentState="{Binding CollectionStateKey,Source={x:Reference self}}" VerticalOptions="FillAndExpand">
+                <StackLayout toolkit:StateContainer.CurrentState="{Binding CollectionStateKey,Source={x:Reference self}}"
+                             VerticalOptions="FillAndExpand">
                     <toolkit:StateContainer.StateViews>
 
-                        <Label toolkit:StateView.StateKey="Empty" Margin="0,0,0,30" Text="{x:Static strings:LocalizedStrings.MultipleChoicePopup_Empty}" Style="{StaticResource CenteredLabel}"></Label>
+                        <Label toolkit:StateView.StateKey="Empty" Margin="0,0,0,30" 
+                               Text="{x:Static strings:LocalizedStrings.MultipleChoicePopup_Empty}" 
+                               Style="{StaticResource CenteredLabel}"></Label>
 
-                        <Label toolkit:StateView.StateKey="LoadingError" Margin="0,0,0,30" Text="{x:Static strings:LocalizedStrings.Errors_Error}" Style="{StaticResource CenteredLabel}"></Label>
+                        <Label toolkit:StateView.StateKey="LoadingError" Margin="0,0,0,30" 
+                               Text="{x:Static strings:LocalizedStrings.Errors_Error}" 
+                               Style="{StaticResource CenteredLabel}"></Label>
 
-                        <Label toolkit:StateView.StateKey="ConnectionError" Margin="0,0,0,30" Text="{x:Static strings:LocalizedStrings.Errors_USOSAPIConnectionError}" Style="{StaticResource CenteredLabel}"></Label>
+                        <Label toolkit:StateView.StateKey="ConnectionError" Margin="0,0,0,30" 
+                               Text="{x:Static strings:LocalizedStrings.Errors_USOSAPIConnectionError}" 
+                               Style="{StaticResource CenteredLabel}"></Label>
 
                         <ActivityIndicator toolkit:StateView.StateKey="Loading" Margin="0,0,0,30"/>
 
@@ -33,7 +42,9 @@
                             </CollectionView.Footer>
                             <CollectionView.ItemTemplate>
                                 <DataTemplate>
-                                    <Button Text="{Binding .}" HeightRequest="100" FontSize="15" BackgroundColor="{StaticResource Primary}" Margin="10" Clicked="OptionsButton_Clicked"/>
+                                    <Button Text="{Binding .}" HeightRequest="75" FontSize="15" 
+                                            BackgroundColor="{StaticResource Primary}" Margin="10" 
+                                            Clicked="OptionsButton_Clicked"/>
                                 </DataTemplate>
                             </CollectionView.ItemTemplate>
                         </CollectionView>
